feat(client): use squiz v1 theme for the scoreboard
#2
