+++ packages/client/src/components/ScoreBoard/ScoreBoard.tsx
@@ -1,36 +1,45 @@
-import { useTheme } from '@react-navigation/native';
-import React from 'react';
+import React, { useEffect } from 'react';
 import { View, StyleSheet } from 'react-native';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 
-import { useScreenWidth } from '../../utils/hooks/screenWidth';
+import playerInfoState, { Player } from '../../global/playerInfoState';
+import socketState from '../../global/socket';
+import { useSocketListener } from '../../utils/hooks/socketListener';
+import Text from '../Text';
 import ScoreBoardContent from './ScoreBoardContent';
 
 export default function ScoreBoard() {
-  const isLargeScreen = useScreenWidth();
-  const { colors } = useTheme();
-  let boardStyle;
+  const players: Player[] = useSocketListener('players', []);
+  const socket = useRecoilValue(socketState);
+  const setPlayerInfo = useSetRecoilState(playerInfoState);
 
-  if (isLargeScreen) {
-    boardStyle = { minWidth: 350, maxWidth: 350 };
-  } else {
-    boardStyle = { flex: 1 };
-  }
+  useEffect(() => {
+    if (!socket || players.length === 0) return;
+    const playerInfos = players.find((player) => player.id === socket.id);
+    if (playerInfos) {
+      setPlayerInfo(playerInfos);
+    }
+  }, [players, socket]);
+
   return (
-    <View
-      style={[
-        styles.container,
-        {
-          backgroundColor: colors.primary,
-        },
-        boardStyle,
-      ]}>
-      <ScoreBoardContent />
-    </View>
+    <>
+      <View style={styles.top}>
+        <Text fontFamily="title" fontSize="xxl">
+          JOUEURS
+        </Text>
+        <Text fontFamily="title" fontSize="xxl">
+          {players.length}
+        </Text>
+      </View>
+      <ScoreBoardContent players={players} />
+    </>
   );
 }
 
 const styles = StyleSheet.create({
-  container: {
-    padding: 4,
+  top: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    paddingBottom: 10,
   },
 });
